+++ src/components/Options/NavOutlet.tsx
@@ -1,10 +1,10 @@
 import { Redirect, Route, Switch } from "react-router-dom";
-import { FrequentlyUsed } from "./pages/tags/FrequentlyUsed";
-import { Settings } from "./pages/Settings";
-import { Sets } from "./pages/tags/sets/Sets";
-import { Users } from "./pages/Users";
-import { Backup } from "./pages/Backup";
-import { About } from "./pages/About";
+import { FrequentlyUsed } from "./Tags/FrequentlyUsed";
+import { Settings } from "./Settings/Settings";
+import { Sets } from "./Tags/Sets/Sets";
+import { Users } from "./Users/Users";
+import { Backup } from "./Backup/Backup";
+import { About } from "./About/About";
 import * as React from "react";
 
 export const NavOutlet: React.FunctionComponent = () => (