+++ src/components/Options/Settings/Settings.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
-import {useSetDocumentTitle} from "../../hooks/setDocumentTitle";
-import {i18n} from "../../constants/i18n";
+import {useSetDocumentTitle} from "../../../hooks/setDocumentTitle";
+import {i18n} from "../../../constants/i18n";
 import {Container, Row, Col} from "react-bootstrap";
 import {useCallback} from "react";
 import {ChangeEvent} from "react";
 import {useEffect} from "react";
-import {getOptions, setTagListLimit, setFrequentlyUsedLimit, setOptions} from "../../slices/options";
+import {getOptions, setTagListLimit, setFrequentlyUsedLimit, setOptions} from "../../../slices/options";
 import {useDispatch, useSelector} from "react-redux";
 import {throttle} from "lodash";
-import {RootState} from "../../common/rootReducer";
+import {RootState} from "../../../common/rootReducer";
 
 export const Settings: React.FunctionComponent = () => {
     const dispatch = useDispatch();