+++ src/components/Options/Tags/Sets/SetView.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { Set, Tag } from "../../../../common/types";
-import TagButton from "../../../../components/TagButton";
+import TagButton from "../../../TagButton";
 import { useCallback, useState } from "react";
 import { TagView } from "./TagView";
 import { Container, Col, Row, Button } from "react-bootstrap";
@@ -8,7 +8,7 @@ import { i18n } from "../../../../constants/i18n";
 import { ChangeEvent } from "react";
 import { useDispatch } from "react-redux";
 import { addTag, updateIconUrl, updateSetName, deleteTag, deleteSet, setSets } from "../../../../slices/sets";
-import { ConfirmModal } from "../../../ConfirmModal";
+import { ConfirmModal } from "../../ConfirmModal";
 import { useHistory } from "react-router-dom";