+++ src/components/Options/Tags/Sets/TagView.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import Tag from "../../../../components/Tag";
+import Tag from "../../../Tag";
 import { Tag as TagType } from "../../../../common/types";
 import { Container, Col, Row } from "react-bootstrap";
-import { ColorPicker } from "../../../ColorPicker";
-import { Badge } from "../../../Badge";
+import { ColorPicker } from "../../ColorPicker";
+import { Badge } from "../../Badge";
 import { i18n } from "../../../../constants/i18n";
 import { useDispatch, useSelector } from "react-redux";
 import { setSets, removeAlias, updateTagColor, updateTagName, updateTagURI, tagSelector } from "../../../../slices/sets";